create to_envp function for execve
