+++ execution/functions/ft_strappend.c
@@ -6,7 +6,7 @@
 /*   By: aes-salm <aes-salm@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/08 12:23:57 by aes-salm          #+#    #+#             */
-/*   Updated: 2021/10/01 11:56:57 by aes-salm         ###   ########.fr       */
+/*   Updated: 2021/10/04 12:59:35 by aes-salm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
